Extract VOPS listing to its own object
