+++ muikku-core-plugins/src/main/java/fi/otavanopisto/muikku/plugins/transcriptofrecords/rest/TranscriptofRecordsRESTService.java
@@ -4,11 +4,9 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import javax.ejb.Stateful;
@@ -28,29 +26,20 @@ import javax.ws.rs.core.StreamingOutput;
 
 import org.apache.commons.lang3.StringUtils;
 import org.codehaus.jackson.map.ObjectMapper;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Entities.EscapeMode;
-import org.jsoup.safety.Cleaner;
-import org.jsoup.safety.Whitelist;
 
 import fi.otavanopisto.muikku.controller.PermissionController;
 import fi.otavanopisto.muikku.controller.PluginSettingsController;
-import fi.otavanopisto.muikku.model.security.Permission;
 import fi.otavanopisto.muikku.model.users.EnvironmentRoleArchetype;
 import fi.otavanopisto.muikku.model.users.EnvironmentRoleEntity;
 import fi.otavanopisto.muikku.model.users.EnvironmentUser;
 import fi.otavanopisto.muikku.model.users.UserEntity;
-import fi.otavanopisto.muikku.model.users.UserGroupEntity;
-import fi.otavanopisto.muikku.model.workspace.WorkspaceEntity;
-import fi.otavanopisto.muikku.model.workspace.WorkspaceUserEntity;
 import fi.otavanopisto.muikku.plugin.PluginRESTService;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.StudiesViewCourseChoiceController;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.TranscriptOfRecordsController;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.TranscriptOfRecordsFileController;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.TranscriptofRecordsPermissions;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.TranscriptofRecordsUserProperties;
-import fi.otavanopisto.muikku.plugins.transcriptofrecords.VopsWorkspace;
+import fi.otavanopisto.muikku.plugins.transcriptofrecords.VopsLister;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.model.StudiesViewCourseChoice;
 import fi.otavanopisto.muikku.plugins.transcriptofrecords.model.TranscriptOfRecordsFile;
 import fi.otavanopisto.muikku.schooldata.CourseMetaController;
@@ -58,14 +47,10 @@ import fi.otavanopisto.muikku.schooldata.GradingController;
 import fi.otavanopisto.muikku.schooldata.RestCatchSchoolDataExceptions;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
 import fi.otavanopisto.muikku.schooldata.WorkspaceController;
-import fi.otavanopisto.muikku.schooldata.entity.GradingScale;
-import fi.otavanopisto.muikku.schooldata.entity.GradingScaleItem;
-import fi.otavanopisto.muikku.schooldata.entity.Optionality;
 import fi.otavanopisto.muikku.schooldata.entity.Subject;
 import fi.otavanopisto.muikku.schooldata.entity.TransferCredit;
 import fi.otavanopisto.muikku.schooldata.entity.User;
 import fi.otavanopisto.muikku.schooldata.entity.WorkspaceAssessment;
-import fi.otavanopisto.muikku.security.MuikkuPermissions;
 import fi.otavanopisto.muikku.session.SessionController;
 import fi.otavanopisto.muikku.users.EnvironmentUserController;
 import fi.otavanopisto.muikku.users.UserController;
@@ -83,7 +68,6 @@ import fi.otavanopisto.security.rest.RESTPermit.Handling;
 public class TranscriptofRecordsRESTService extends PluginRESTService {
 
   private static final long serialVersionUID = -6752333351301485518L;
-  private static final int MAX_COURSE_NUMBER = 15;
 
   @Inject
   private TranscriptOfRecordsFileController transcriptOfRecordsFileController;
@@ -130,16 +114,6 @@ public class TranscriptofRecordsRESTService extends PluginRESTService {
   @Inject
   private Logger logger;
   
-  private String clean(String html) {
-    if (html == null) {
-      return null;
-    }
-    Document doc = Jsoup.parseBodyFragment(html);
-    doc = new Cleaner(Whitelist.none()).clean(doc);
-    doc.outputSettings().escapeMode(EscapeMode.xhtml);
-    return doc.body().html();
-  }
-
   @GET
   @Path("/files/{ID}/content")
   @RESTPermit(handling = Handling.INLINE)
@@ -241,241 +215,36 @@ public class TranscriptofRecordsRESTService extends PluginRESTService {
       }
     });
     
-    for (Subject subject : subjects) {
-      boolean subjectHasCourses = false;
-      if (vopsController.subjectAppliesToStudent(student, subject)) {
-        List<VopsRESTModel.VopsEntry> entries = new ArrayList<>();
-        course: for (int courseNumber=1; courseNumber<MAX_COURSE_NUMBER; courseNumber++) {
-          boolean hasTransferCredit = false;
+    VopsLister lister = new VopsLister(
+      subjects,
+      vopsController,
+      student,
+      transferCredits,
+      curriculumIdentifier,
+      workspaceController,
+      workspaceUserEntityController,
+      studentIdentifier,
+      studentAssessments,
+      userGroupEntityController,
+      permissionController,
+      studiesViewCourseChoiceController,
+      studentIdentifierString,
+      gradingController,
+      educationTypeMapping
+    );
+    
+    lister.performListing();
 
-          for (TransferCredit transferCredit : transferCredits) {
-            boolean subjectsMatch = Objects.equals(
-                transferCredit.getSubjectIdentifier(),
-                new SchoolDataIdentifier(subject.getIdentifier(), subject.getSchoolDataSource()));
-            boolean courseNumbersMatch = Objects.equals(
-                transferCredit.getCourseNumber(),
-                courseNumber);
-            if (subjectsMatch && courseNumbersMatch) {
-              String grade = "";
-              GradingScaleItem gradingScaleItem = null;
-              Mandatority mandatority = Mandatority.MANDATORY;
-              if (transferCredit.getOptionality() == Optionality.OPTIONAL) {
-                mandatority = Mandatority.UNSPECIFIED_OPTIONAL;
-              }
-
-              if (transferCredit.getGradeIdentifier() != null
-                  && transferCredit.getGradingScaleIdentifier() != null) {
-                gradingScaleItem = findGradingScaleItemCached(
-                    transferCredit.getGradingScaleIdentifier(),
-                    transferCredit.getGradeIdentifier()
-                );
-                
-                String gradeName = gradingScaleItem.getName();
-                if (!StringUtils.isBlank(gradeName)) {
-                  if (gradeName.length() > 2)
-                    grade = gradeName.substring(0, 2);
-                  else
-                    grade = gradeName;
-                }
-              }
-              entries.add(new VopsRESTModel.VopsItem(
-                  courseNumber,
-                  CourseCompletionState.ASSESSED,
-                  (String)null,
-                  mandatority,
-                  grade,
-                  false,
-                  transferCredit.getCourseName(),
-                  ""
-              ));
-              hasTransferCredit = true;
-              break;
-            }
-          }
-
-          List<VopsWorkspace> workspaces =
-              vopsController.listWorkspaceIdentifiersBySubjectIdentifierAndCourseNumber(
-                  subject.getSchoolDataSource(),
-                  subject.getIdentifier(),
-                  courseNumber);
-          
-          List<WorkspaceAssessment> workspaceAssessments = new ArrayList<>();
-          
-          boolean correctCurriculum = false;
-          
-          if (curriculumIdentifier == null) {
-            correctCurriculum = true;
-          } else {
-            for (VopsWorkspace workspace : workspaces) {
-              if (workspace.getCurriculumIdentifiers().contains(curriculumIdentifier)) {
-                correctCurriculum = true;
-                break;
-              }
-            }
-          }
-
-          if (!hasTransferCredit && !workspaces.isEmpty() && correctCurriculum) {
-            SchoolDataIdentifier educationSubtypeIdentifier = null;
-            boolean workspaceUserExists = false;
-            String name = "";
-            String description = "";
-            boolean canSignUp = false;
-
-            for (VopsWorkspace workspace : workspaces) {
-              WorkspaceEntity workspaceEntity =
-                  workspaceController.findWorkspaceEntityById(workspace.getWorkspaceIdentifier());
-              WorkspaceUserEntity workspaceUser =
-                  workspaceUserEntityController.findWorkspaceUserByWorkspaceEntityAndUserIdentifier(
-                      workspaceEntity,
-                      studentIdentifier);
-              WorkspaceAssessment workspaceAssesment = studentAssessments.get(workspace.getWorkspaceIdentifier());
-              
-              List<UserGroupEntity> userGroupEntities = userGroupEntityController.listUserGroupsByUserIdentifier(studentIdentifier);
-              
-              Permission permission = permissionController.findByName(MuikkuPermissions.WORKSPACE_SIGNUP);
-              for (UserGroupEntity userGroupEntity : userGroupEntities) {
-                if (permissionController.hasWorkspaceGroupPermission(workspaceEntity, userGroupEntity, permission)) {
-                  canSignUp = true;
-                  break;
-                }
-              }
-              
-              if (workspaceAssesment != null) {
-                workspaceAssessments.add(workspaceAssesment);
-              }
-
-              if (workspaceUser != null) {
-                workspaceUserExists = true;
-              }
-            }
-
-            if (!canSignUp) {
-              entries.add(new VopsRESTModel.VopsPlaceholder());
-              continue course;
-            }
-            
-            for (VopsWorkspace workspace : workspaces) {
-              name = workspace.getName();
-              if (name != null) {
-                break;
-              }
-            }
-            
-            for (VopsWorkspace workspace : workspaces) {
-              description = workspace.getDescription();
-              if (description != null) {
-                break;
-              }
-            }
-            
-            for (VopsWorkspace workspace : workspaces) {
-              educationSubtypeIdentifier = workspace.getEducationSubtypeIdentifier();
-              if (educationSubtypeIdentifier != null) {
-                break;
-              }
-            }
-            
-            Mandatority mandatority = educationTypeMapping.getMandatority(educationSubtypeIdentifier);
-            CourseCompletionState state = CourseCompletionState.NOT_ENROLLED;
-            String grade = null;
-            if (workspaceUserExists) {
-              state = CourseCompletionState.ENROLLED;
-            }
-            for (WorkspaceAssessment workspaceAssessment : workspaceAssessments) {
-              if (!workspaceAssessment.getPassing()) {
-                state = CourseCompletionState.FAILED;
-                break;
-              }
-            }
-            for (WorkspaceAssessment workspaceAssessment : workspaceAssessments) {
-              if (workspaceAssessment.getPassing()) {
-                state = CourseCompletionState.ASSESSED;
-                numCourses++;
-                if (mandatority == Mandatority.MANDATORY) {
-                  numMandatoryCourses++;
-                }
-
-                SchoolDataIdentifier gradingScaleIdentifier = workspaceAssessment.getGradingScaleIdentifier();
-                if (gradingScaleIdentifier == null) {
-                  break;
-                }
-                SchoolDataIdentifier gradeIdentifier = workspaceAssessment.getGradeIdentifier();
-                if (gradeIdentifier == null) {
-                  break;
-                }
-                GradingScaleItem gradingScaleItem = findGradingScaleItemCached(gradingScaleIdentifier, gradeIdentifier);
-                String gradeName = gradingScaleItem.getName();
-                if (!StringUtils.isBlank(gradeName)) {
-                  // 2 characters is enough to cover cases like "10" and "Suoritettu/Saanut opetusta" unambiguously
-                  // and still looking good in the matrix
-                  
-                  if (gradeName.length() > 2)
-                    grade = gradeName.substring(0, 2);
-                  else
-                    grade = gradeName;
-                }
-                
-                break;
-              }
-            }
-            
-            StudiesViewCourseChoice courseChoice = studiesViewCourseChoiceController.find(
-                new SchoolDataIdentifier(subject.getIdentifier(), subject.getSchoolDataSource()).toId(),
-                courseNumber,
-                studentIdentifierString);
-            if (state == CourseCompletionState.NOT_ENROLLED
-                && courseChoice != null) {
-              state = CourseCompletionState.PLANNED;
-            }
-            
-            entries.add(new VopsRESTModel.VopsItem(
-                courseNumber,
-                state,
-                educationSubtypeIdentifier != null ? educationSubtypeIdentifier.toId() : null,
-                mandatority,
-                grade,
-                workspaceUserExists,
-                clean(name),
-                clean(description)
-            ));
-            subjectHasCourses = true;
-          } else if (!hasTransferCredit) {
-            entries.add(new VopsRESTModel.VopsPlaceholder());
-          }
-        }
-        if (subjectHasCourses) {
-          rows.add(new VopsRESTModel.VopsRow(
-              subject.getCode(),
-              new SchoolDataIdentifier(subject.getIdentifier(), subject.getSchoolDataSource()).toId(),
-              entries));
-        }
-      }
-    }
-
-    VopsRESTModel result = new VopsRESTModel(rows, numCourses, numMandatoryCourses, true);
+    VopsRESTModel result = new VopsRESTModel(
+        lister.getRows(),
+        lister.getNumCourses(),
+        lister.getNumMandatoryCourses(),
+        lister.isOptedIn()
+    );
 
     return Response.ok(result).build();
   }
   
-  
-  private GradingScaleItem findGradingScaleItemCached(SchoolDataIdentifier gradingScaleIdentifier, SchoolDataIdentifier gradingScaleItemIdentifier) {
-    GradingScaleItemCoordinates key = new GradingScaleItemCoordinates(gradingScaleIdentifier, gradingScaleItemIdentifier);
-    if (!gradingScaleCache.containsKey(key)) {
-      GradingScale gradingScale = gradingController.findGradingScale(gradingScaleIdentifier);
-      if (gradingScale == null) {
-        logger.log(Level.SEVERE, "Grading scale not found for identifier: %s", gradingScaleIdentifier);
-        return null;
-      }
-      for (GradingScaleItem gradingScaleItem : gradingController.listGradingScaleItems(gradingScale)) {
-        GradingScaleItemCoordinates newItemKey = new GradingScaleItemCoordinates(
-            gradingScaleIdentifier,
-            new SchoolDataIdentifier(gradingScaleItem.getIdentifier(), gradingScaleItem.getSchoolDataSource()));
-        gradingScaleCache.put(newItemKey, gradingScaleItem);
-      }
-    }
-    return gradingScaleCache.get(key);
-  }
-
   private HopsRESTModel createHopsRESTModelForStudent(SchoolDataIdentifier userIdentifier) {
     User user = userController.findUserByIdentifier(userIdentifier);
     UserEntity userEntity = userEntityController.findUserEntityByUser(user);
@@ -657,34 +426,4 @@ public class TranscriptofRecordsRESTService extends PluginRESTService {
     return Response.ok().entity(model).build();
   }
   
-  private static class GradingScaleItemCoordinates {
-    public GradingScaleItemCoordinates(
-        SchoolDataIdentifier gradingScaleIdentifier,
-        SchoolDataIdentifier gradingScaleItemIdentifier
-    ) {
-      this.gradingScaleIdentifier = gradingScaleIdentifier;
-      this.gradingScaleItemIdentifier = gradingScaleItemIdentifier;
-    }
-    
-    @Override
-    public boolean equals(Object obj) {
-      if (obj instanceof GradingScaleItemCoordinates) {
-        GradingScaleItemCoordinates coords = (GradingScaleItemCoordinates) obj;
-        return coords.gradingScaleIdentifier.equals(this.gradingScaleIdentifier) &&
-               coords.gradingScaleItemIdentifier.equals(this.gradingScaleItemIdentifier);
-      } else {
-        return false;
-      }
-    }
-    
-    @Override
-    public int hashCode() {
-      return gradingScaleIdentifier.hashCode() * 37 + gradingScaleItemIdentifier.hashCode();
-    }
-    
-    private final SchoolDataIdentifier gradingScaleIdentifier;
-    private final SchoolDataIdentifier gradingScaleItemIdentifier;
-  }
-  
-  private Map<GradingScaleItemCoordinates, GradingScaleItem> gradingScaleCache = new HashMap<>();
 }
